API

InvariantManifolds.OneDManifoldProblem — Type

OneDManifoldProblem{F,T}

OneDManifoldProblem is a struct to contain the main information for continuing the one-dimensional manifold of a nonlinear map.

Fields

f the nonlinear map, which should has the form f(x,p) and return a SVector;

para the parameters of the nonlinear map;

amax the maximum angle between points when continuing the manifold;

d the maximum distance between points when continuing the manifold;

dsmin the minimum arc length allowing; note that if in a continuation point, this value is achieved and the angle as well as the distance values are not achieved, then we will record this point as a FlawPoint.

Convenient consturctors are OneDManifoldProblem(f) and OneDManifoldProblem(f,para)

InvariantManifolds.OneDManifold — Type

OneDManifold{F,S,N,T}

OneDManifold is a struct contains all the information of the one-dimensional numerical manifold.

Fields

prob the problem OneDManifoldProblem;

data the numerical data that should be Vector{Vector{S}}, where S is the interpolation curve (we use DataInterpolation in this package);

flawpoints the flaw points generated during continuation.

InvariantManifolds.NSOneDManifoldProblem — Type

NSOneDManifoldProblem{F,T}

NSOneDManifoldProblem is a struct to contain the main information for continuing the non-smooth one-dimensional manifold of the time-T-map of a non-smooth ODE.

Fields

f the struct NSSetUp, see NSSetUp;

para the parameters of the nonlinear map;

amax the maximum angle between points when continuing the manifold;

d the maximum distance between points when continuing the manifold;

ϵ the max value of the following expression: $\max\{|H(x_0,T)|,H(x_1,T)\},$ where $H(x,t)$ is the hypersurface the manifold cross, $x_0$ and $x_1$ are points before and after the cross, $T$ is the end of the time-$T$-map (from 0 to $T$).

dsmin the minimum arc length allowing; note that if in a continuation point, this value is achieved and the angle as well as the distance values are not achieved, then we will record this point as a FlawPoint.

Convenient consturctors are NSOneDManifoldProblem(f) and NSOneDManifoldProblem(f,para)

InvariantManifolds.NSOneDManifold — Type

NSOneDManifold{F,S,N,T}

NSOneDManifold is a struct contains all the information of the non-smooth one-dimensional numerical manifold.

Fields

prob the problem OneDManifoldProblem;

data the numerical data that should be Vector{Vector{Vector{S}}}, where S is the interpolation curve (we use DataInterpolation in this package);

flawpoints the flaw points generated during continuation.

InvariantManifolds.TwoDManifoldProblem — Type

TwoDManifoldProblem{F,T}

TwoDManifoldProblem is a struct to contain the main information for continuing the two-dimensional manifold of a nonlinear map.

Fields

f the nonlinear map, which should has the form f(x,p) and return a SVector;

para the parameters of the nonlinear map;

amax the maximum angle between points when continuing the manifold;

d the maximum distance between points when continuing the manifold;

dcircle the maximum distance between circles when continuing the manifold;

dsmin the minimum arc length allowing; note that if in a continuation point, this value is achieved and the angle as well as the distance values are not achieved, then we will record this point as a FlawPoint.

Convenient consturctors are TwoDManifoldProblem(f) and TwoDManifoldProblem(f,para)

InvariantManifolds.TwoDManifold — Type

TwoDManifold{F,S,N,T}

TwoDManifold is a struct contains all the information of the two-dimensional numerical manifold of a nonlinear map.

Fields

prob the problem TwoDManifoldProblem;

data the numerical data that should be Vector{S}, where S is the interpolation curve (we use DataInterpolation in this package);

flawpoints the flaw points generated during continuation.

InvariantManifolds.VTwoDManifoldProblem — Type

VTwoDManifoldProblem{F,T}

VTwoDManifoldProblem is a struct to contain the main information for continuing the two-dimensional manifold of an autonomous vector field.

Fields

f the time flow map of the vector field, which should has the form f(x,p) and return a SVector; note that the vector field should be regularized, e.g., $\dot{x}=v(x)$ should be rewritten as $\dot{x}=v(x)/|v(x)|$;

para the parameters of the time flow map;

amax the maximum angle between points when continuing the manifold;

d the maximum distance between points when continuing the manifold;

dsmin the minimum arc length allowing; note that if in a continuation point, this value is achieved and the angle as well as the distance values are not achieved, then we will record this point as a FlawPoint.

Convenient consturctors are VTwoDManifoldProblem(f) and VTwoDManifoldProblem(f,para)

InvariantManifolds.VTwoDManifold — Type

VTwoDManifold{F,S,N,T}

VTwoDManifold is a struct contains all the information of the two-dimensional numerical manifold of an autonomous vector field.

Fields

prob the problem VTwoDManifoldProblem;

data the numerical data that should be Vector{S}, where S is the interpolation curve (we use DataInterpolation in this package);

flawpoints the flaw points generated during continuation.

InvariantManifolds.NSState — Type

NSState{N,T<:Number} <: AbstractVector{T}

The struct NSState is to record the events data for a time-T-map.

Fields

state the final state of the time-T-map;

event_at is a integer vector that contains the history of the events happened.

The meaning of event_at in PiecewiseV and BilliardV systems is quite clear. For a simple Fillippov system SFilippovV, we record three events:

1 represents the event that flow enters the sliding surface from H<0 or H>0;

2 represents the event that flow cross the sliding surface;

3 represents the event that flow slides out the sliding surface from H=0 to H<0 or H>0.

InvariantManifolds.NSSolution — Type

NSSolution{N,T<:Number}

The NSSolution is a struct to contain all information of the solution of a non-smooth ODE system.

Fields

sol ODESolution solved by OrdinaryDiffEq;

event_t the times when events happen;

event_state the solution's state when events happen;

event_at is a vector that contains integers indicating which event happen.

InvariantManifolds.NSSetUp — Type

NSSetUp{T}

NSSetUp is a struct to contain all the information needed in continuing the manifold of non-smooth ODE.

Fields

f the Non-smooth vector field, like PiecewiseV;

timespan the time span of time-T-map;

timetmap the time-t-map of non-smooth ODE, which maps a NSState and parameters of ODE to a NSState.

InvariantManifolds.setmap — Function

setmap(v::T, timespan, alg, N, T; region_detect=_region_detect, extra...)

The function setmap is to get a NSSetUp.

Parameters

v a nonsmooth vector field like PiecewiseV or BilliardV.

timespan the time span of the time-T-map.

alg algorithm in OrdinaryDiffEq to solve ODE.

N the dimension of the vector field.

T number type used in computation. Usually, Float64 is enough.

To ensure type stable, the numbers in timespan should be type of T. The last two parameters has to be specified, since we have to store the event data.

Keyword arguments

For ContinuousVectorField such as PiecewiseV, we have keyword argument:

region_detect=_region_detect the region detect function to determine which domain the state in.

You can also pass the keywords of solve of OrdinaryDiffEq to this function, except the callback and saving related keywords.

InvariantManifolds.timetmap — Function

timetmap(v::T, para, timespan, alg; region_detect=_region_detect, extra...)

The function timetmap is similar to setmap. The output of this function is a function which maps a SVector and parameters of ODE to a SVector, i.e. the time-T-map.

Parameters

v a ContinuousVectorField or JumpVectorField like PiecewiseV or BilliardV.

timespan the time span of the time-T-map. To ensure type stable, the numbers timespan should be type of float numbers you are using.

alg algorithm in OrdinaryDiffEq to solve ODE.

Keyword arguments

For ContinuousVectorField such as PiecewiseV, there is a keyword argument:

region_detect=_region_detect the region detect function to determine which domain the state in.

You can also pass the keywords of solve of OrdinaryDiffEq to this function, except the callback and saving related keywords.

InvariantManifolds.ns_solver — Function

ns_solver(v::T, para, timespan, alg, N, T)

The function ns_solver is similar to timetmap. The output of this function is a function which maps a SVector to a NSSolution. This NSSolution contain all data of an non-smooth ODE solution.

Parameters

v a ContinuousVectorField or JumpVectorField like PiecewiseV or BilliardV.

para the parameter of the vector field.

timespan the time span of the time-T-map.

alg algorithm in OrdinaryDiffEq to solve ODE.

N the dimension of the vector field.

T number type used in computation.

To ensure type stable, the numbers in para and timespan should be type of T. The last two parameters have to be specified, since we need to store the event data. You can also pass the keywords of solve of OrdinaryDiffEq to this function, except the callback and saving related keywords.

InvariantManifolds.PiecewiseV — Type

PiecewiseV <: ContinuousVectorField

Piecewise smooth vector field.

Fields

fs is a vector of smooth vector fields in different regions.

regions is a vector of the region functions: [r1,r2,...], where r1(x,p,t) should return a Bool value to indicate that x is in this region or not.

hypers is a vector of the hyper surfaces separating the regions.

n is a integer to switch between vector fields. It can be set to any integer when construct a PiecewiseV.

Example

using StaticArrays, InvariantManifolds
f1(x,p,t)=SA[x[2],-2x[1]]
f2(x,p,t)=SA[x[2],-x[1]]
dom1(x,p,t)=x[1]>0
dom2(x,p,t)=x[2]<0
hyper(x,p,t)=x[1]
PiecewiseV((f1,f2),(dom1,dom2),(hyper,))

The above codes generate a piecewise smooth vector field, which when x[1]>0 is f1, and when x[2]<0 is f2. The hyper surface separating these smooth vector fields is x[1]=0.

InvariantManifolds.BilliardV — Type

BilliardV <: JumpVectorField

A vector field with multiple hyper surfaces such that the flow jump when hits these hyper surfaces.

Fields

f is the vector field, of type f(x,p,t), and its output is a SVector;

hypers is tuple of hyper surfaces:(h1,h2,...), h1(x,p,t);

irules is tuple of rules on hyper surfaces:(r1,r2,r3,...).

InvariantManifolds.SFilippovV — Type

SFilippovV <: ContinuousVectorField

SFilippovV means simple Fillippov vector fields, which means that there only exists one hyper surface to separate the phase space.

Fields

fs vector fields in two sides of hyper surface. The slide vector field can be generated automatically.

hyper hyper surface.

dhyper grad of hyper surface. Warn!!! The grad must point to the second of fs.

exit conditions to exit the hyper surface, which can also be generated automatically.

n is a integer to switch between vector fields. It can be set to any integer when construct a SFilippovV.

InvariantManifolds.FlawPoint — Type

FlawPoint{N,T}

FlawPoint is a struct to record the points that don't satisfy the angle and distance request while the program has reached the minimum arc length.

Fields

point flaw point in the process of continuation;

α the angle recorded;

d the distance recorded.

InvariantManifolds.gen_segment — Function

segment(saddle, direction)

Generating n points at saddle in the direction, with length d, with default n=150 and d=0.01. Another Convenient consturctor is segment(p::Saddle).

InvariantManifolds.gen_disk — Function

gen_disk(p, times)

gen_disk is a function to generate circles around the saddle, which represented as the local manifold.

Parameters

p the struct Saddle which should contains two unstable directions; the complex eigenvalues and eigenvectors are allowed.

times the iteration times; this parameter is needed to adjust the torsion in different directions in the process of continuation.

Keyword argument

n the number of point in each circle, default to be 150;

d the max distance between points in a single circle, default to be 0.002;

r the size of the disk, default to be 0.05;

circles the number of the circles, default to be 10.

InvariantManifolds.initialize — Function

initialize(prob, points)

This is a function to initialize the continuation process. Its output is a manifold struct.

Parameters

prob the problem such as OneDManifoldProblem.

points the points in the local manifold. For one dimensional manifolds, these points should be a Vector{SVector} and the start point should be the saddle. For two dimensional manifolds, these points should be a Vector{Vector{S}} and its first element should like [saddle, saddle, saddle]. Note that in the both cases, the functions gen_segment, gen_disk, and gen_circle can generate these points easily.

Keyword argument

interp the interpolation method used, default to be LinearInterpolation.

InvariantManifolds.grow! — Function

grow!(manifold)

One time iteration to grow the manifold.

Parameters

manifold the manifold struct.

Keyword argument

interp the interpolation method used, default to be LinearInterpolation.

InvariantManifolds.growmanifold — Function

growmanifold(prob, points, N)

This is the mani function to continuate the numerical manifolds. Its output is a manifold struct.

Parameters

prob the problem such as OneDManifoldProblem.

points the points in the local manifold. For one dimensional manifolds, these points should be a Vector{SVector} and the start point should be the saddle. For two dimensional manifolds, these points should be a Vector{Vector{S}} and its first element should like [saddle, saddle, saddle]. Note that in the both cases, the functions segment and disk can generate these points easily.

N the number of iterations.

Keyword argument

interp the interpolation method used, default to be LinearInterpolation.

InvariantManifolds.Saddle — Type

Saddle{N,T,S}

Saddle is a struct to contain the information of a saddle point needed in continuing the manifold of non-smooth ODE. For an ODE's saddle, this struct can be constructed by the function findsaddle.

Fields

saddle the location of the saddle point;

unstable_directions the unstable directions;

unstable_eigen_values eigenvalues of the linearized map in the saddle at the unstable eigenvectors.

InvariantManifolds.findsaddle — Function

findsaddle(v, dv, timespan, x, p)

findsaddle is a function to find the saddle of the time-T-map of smooth ODE systems, by using the Newton's method.

Parameters

v the vector field, which should be the form f(x,p,t) and return a SVector;

dv the Jacobi matrix function of v, which should should be the form dv(x,p,t) and return a SMatrix;

timespan the time span of the time-T-map;

x the initial point to iterate.

Keyword arguments

n maximum iterate times, default to be 100;

abstol absolute tolerance for the fixed point, default to be 1e-8;

alg the algorithm used to solve the ODE, default to be Tsit5().