API

InvariantManifolds.segment — Function

segment(point, direction, n, d)

Generating n points at point in the direction, with length d

InvariantManifolds.addpoints — Function

InvariantManifolds.addpoints(f, p, ic1::IterationCurve{N,T}, min)

addpoints will add enough points from ic1 so that its image of points are dense, i.e., the distance of nearby points will less than min.

InvariantManifolds.initialise_curve — Function

InvariantManifolds.initialise_curve(points, map, parameters)

This is a function to get the iteration curve. From [p0,...,pn], where p0 is the saddle point, to get [pn,...,f(pn)].

points stands for distributed points in a local manifolds, points[1] is the saddle;

map is the map function;

parameters is the parameters of the map.

InvariantManifolds.initialise_curve(points, tmap)

Initialise the curve of ODE's time-T-map.

InvariantManifolds.PiecewiseV — Type

PiecewiseV

Piecewise smooth vector field. Fields:

fs is a vector of smooth vector fields in different regions.

regions is a vector of the region functions: [r1,r2,...], where r1(x,p,t) should return a Bool value to indicate that x is in this region or not.

hypers is a vector of the hypersurfaces separating the regions.

InvariantManifolds.BilliardV — Type

BilliardV

The struct BilliardV has three fields:

f is the vector field, of type f(x,p,t), and its output is a SVector;

hypers is vector of hypersurfaces:[h1,h2,...], h1(x,p,t);

irules is vector of rules on hypersurfaces:[r1,r2,r3,...]; the rules must be independent of time t, i.e., r1(x,p).

InvariantManifolds.SFilippovV — Type

SFilippovV

SFilippovV means simple Fillippov vector fields, which means that there only exists one hypersurface to separate the phase space. Fields:

fs vector fields in two sides of hypersurface. The slide vector field can be generated automatically.

hyper hypersurface.

dhyper grad of hypersurface. Warn!!! The grad must point to the second of fs.

exit conditions to exit the hypersurface, which can also be generated automatically

InvariantManifolds.NSSetUp — Type

NSSetUp{T}

NSSetUp is a struct to contatin all the information needed in continuing the manifold. Fields:

f::T the Non-smooth vector field, like PiecewiseV;

p the parameter;

timespan the time span of time-T-map;

timetmap the time-t-map of nonsmooth ODE, which maps a State to a NSState;

alg algorithm used to solve ODE.

InvariantManifolds.setmap — Function

setmap(v::PiecewiseV, para, timespan, alg; region_detect=_region_detect, extra...)

Function to get a NSSetUp.