API

InvariantManifolds.State — Type

State{N,T<:Number}

State is a struct contain point data in a numerical manifold.

Fields

state a SVector reprents the point in phase space;

s the parameter of this point.

InvariantManifolds.NSState — Type

NSState{N,T<:Number}

The struct NSState is to record the events data for a time-T-map.

Fields

state the final state of the time-T-map;

event_t the times when events happen;

event_state the solution's state when events happen;

event_at is a vector that contains integers indicating which event happen;

s the parameter of the NSState's preimage.

The meaning of event_at in PiecewiseV and BilliardV systems is quite clear. For a simple Fillippov system SFilippovV, we record three events:

1 reprents the event that flow enters the slding surface from H<0 or H>0;

2 reprents the event that flow cross the slding surface;

3 reprents the event that flow slides out the slding surface from H=0 to H<0 or H>0.

InvariantManifolds.IterationCurve — Type

IterationCurve{N,T<:Number}

IterationCurve is a struct that contains line data in a numerical manifold.

Fields

states a Vector{State} contains the points in phase space;

pcurve the paramtric curve of these states, which is a LinearInterpolation.

InvariantManifolds.NSSolution — Type

NSSolution{N,T<:Number}

The NSSolution is a struct to contain all information of the solution of a nonsmooth ODE system.

Fields

sol ODESolution solved by OrdinaryDiffEq;

event_t the times when events happen;

event_state the solution's state when events happen;

event_at is a vector that contains integers indicating which event happen.

InvariantManifolds.NSSetUp — Type

NSSetUp{T}

NSSetUp is a struct to contatin all the information needed in continuing the manifold of nonsmooth ODE.

Fields

f::T the Non-smooth vector field, like PiecewiseV;

timespan the time span of time-T-map;

timetmap the time-t-map of nonsmooth ODE, which maps a State and parameters of ODE to a NSState.

Warn

In a ContinuousVectorField,this parameters must be long than the real parameter. For example, if the parameter of your system is [0.1,2.0], than you must set it to [0.1,2.0,1.0]. The value of the last parameter is meaninglless, just for switching the vector fields.

InvariantManifolds.AnnulusBoundaries — Type

AnnulusBoundaries

A struct contains datas when generating the two dimensional manifold.

Fields

inner an IterationCurve reprents inner boundary;

outer an IterationCurve reprents outer boundary;

InvariantManifolds.PiecewiseV — Type

PiecewiseV <: ContinuousVectorField

Piecewise smooth vector field.

Fields

fs is a vector of smooth vector fields in different regions.

regions is a vector of the region functions: [r1,r2,...], where r1(x,p,t) should return a Bool value to indicate that x is in this region or not.

hypers is a vector of the hypersurfaces separating the regions.

Example

using StaticArrays, InvariantManifolds
f1(x,p,t)=SA[x[2],-2x[1]]
f2(x,p,t)=SA[x[2],-x[1]]
dom1(x,p,t)=x[1]>0
dom2(x,p,t)=x[2]<0
hyper(x,p,t)=x[1]
PiecewiseV([f1,f2],[dom1,dom2],[hyper])

The above codes generate a piecewise smooth vector field, which when x[1]>0 is f1, and when x[2]<0 is f2. The hypersurface separating these smooth vector fields is x[1]=0.

InvariantManifolds.BilliardV — Type

BilliardV <: JumpVectorField

A vector field with multiple hypersurfaces such that the flow jump when hits these hypersurfaces.

Fields

f is the vector field, of type f(x,p,t), and its output is a SVector;

hypers is vector of hypersurfaces:[h1,h2,...], h1(x,p,t);

irules is vector of rules on hypersurfaces:[r1,r2,r3,...]; the rules must be independent of time t, i.e., r1(x,p).

InvariantManifolds.SFilippovV — Type

SFilippovV <: ContinuousVectorField

SFilippovV means simple Fillippov vector fields, which means that there only exists one hypersurface to separate the phase space.

Fields

fs vector fields in two sides of hypersurface. The slide vector field can be generated automatically.

hyper hypersurface.

dhyper grad of hypersurface. Warn!!! The grad must point to the second of fs.

exit conditions to exit the hypersurface, which can also be generated automatically

InvariantManifolds.setmap — Function

setmap(v::T, timespan, alg, N, T; region_detect=_region_detect, extra...)

The function setmap is to get a NSSetUp.

Parameters

v a ContinuousVectorField or JumpVectorField like PiecewiseV or BilliardV.

timespan the time span of the time-T-map.

alg algorithm in OrdinaryDiffEq to solve ODE.

N the dimension of the vector field.

T number type used in computation. Usually, Float64 is enough.

To ensure type stable, the numbers in timespan should be type of T. The last two parameters has to be specified, since we have to store the event data. You can also pass the keywords of solve of OrdinaryDiffEq to this function, except the callback and saving related keywords.

InvariantManifolds.timetmap — Function

timetmap(v::T, para, timespan, alg; region_detect=_region_detect, extra...)

The function timetmap is similar to setmap. The output of this function is a function which maps a SVector and parameters of ODE to a SVector, i.e. the time-T-map. Warn!!! In a ContinuousVectorField,this parameters must be long than the real parameter. For example, if the parameter of your system is [0.1,2.0], than you must set it to [0.1,2.0,1.0]. The value of the last parameter is meaninglless, just for switching the vector fields.

Parameters

v a ContinuousVectorField or JumpVectorField like PiecewiseV or BilliardV.

timespan the time span of the time-T-map.

alg algorithm in OrdinaryDiffEq to solve ODE.

To ensure type stable, the numbers timespan should be type of T. You can also pass the keywords of solve of OrdinaryDiffEq to this function, except the callback and saving related keywords.

InvariantManifolds.ns_solver — Function

ns_solver(v::T, para, timespan, alg, N, T)

The function ns_solver is similar to timetmap. The output of this function is a function which maps a SVector to a NSSolution. This NSSolution contain all data of an non-smooth ODE solution.

Parameters

v a ContinuousVectorField or JumpVectorField like PiecewiseV or BilliardV.

para the parameter of the vectorfield.

timespan the time span of the time-T-map.

alg algorithm in OrdinaryDiffEq to solve ODE.

N the dimension of the vector field.

T number type used in computation.

To ensure type stable, the numbers in para and timespan should be type of T. The last two parameters has to be specified, since we have to store the event data. You can also pass the keywords of solve of OrdinaryDiffEq to this function, except the callback and saving related keywords.

InvariantManifolds.segment — Function

segment(point, direction, n, d)

Generating n points at point in the direction, with length d

InvariantManifolds.generate_curves — Function

generate_curves(f, p, seg, d, n)

This function is to generate one-dimensional manifold of smooth mapping or a time-T-map of a non-smooth ODE.

Parameters

f a discrete map of type f(x,p)=SA[...] or a NSSetUp of non-smooth ODE;

para the parameters of systems;

seg inintial segment of manifolds, which can be generated by function segment if the unstable direction is known;

d max distence between points;

n iteration times.

InvariantManifolds.generate_surface — Function

generate_surface(f, p, saddle, v1, v2, N, r, δ)

Function to generate the two dimension manifold of a vector field.

Parameters

f the time-T-map a the vector field;

p parameter of the system;

saddle saddle fixed point of the ODE system;

v1 an instable direction of the saddle;

v2 anthor instable direction of the saddle;

N iteration times;

r the raduis of origin disk to extend;

δ the max distance between points when iterating.

Keyword arguments

n=150 the number of points in the boundary of the origin disk to extend.

InvariantManifolds.addpoints — Function

InvariantManifolds.addpoints(f, p, ic1::IterationCurve{N,T}, min)

addpoints will add enough points from ic1 so that its image of points are dense, i.e., the distance of nearby points will less than min. This function will also delete extra points so that the distance of nearby points aren't two small.

InvariantManifolds.initialise_curve — Function

InvariantManifolds.initialise_curve(points, map, parameters)

This is a function to get the iteration curve. From [p0,...,pn], where p0 is the saddle point, to get [pn,...,f(pn)].

Parameters

points stands for distributed points in a local manifolds, points[1] is the saddle;

map is the map function;

parameters is the parameters of the map.

InvariantManifolds.ns_initialise_curve — Function

InvariantManifolds.ns_initialise_curve(points, tmap, para)

Initialise the curve of nonsmooth ODE's time-T-map.